Add Security with Ratchet Rotation + HMAC + JWT + MasterToken + DeviceToken and add smart worker to more speed up request and process data using parallel processing

+++ internal/domain/schemas.go
@@ -8,7 +8,7 @@ import (
 
 // UsersCache représente la structure du cache "users"
 var UsersSchema = map[string]reflect.Kind{
-	"id":                 reflect.Int, // UUID
+	"id":                 reflect.Int64, // UUID
 	"username":           reflect.String,
 	"email":              reflect.String,
 	"email_verified":     reflect.Bool,
@@ -20,7 +20,7 @@ var UsersSchema = map[string]reflect.Kind{
 	"birthdate":          reflect.Struct, // Date au format string
 	"sex":                reflect.Int,
 	"bio":                reflect.String,
-	"profile_picture_id": reflect.Int, // UUID
+	"profile_picture_id": reflect.Int64, // UUID
 	"grade":              reflect.Int,
 	"location":           reflect.String,
 	"school":             reflect.String,
@@ -36,8 +36,8 @@ var UsersSchema = map[string]reflect.Kind{
 
 // UserSettingsCache représente la structure du cache "user_settings"
 var UserSettingsSchema = map[string]reflect.Kind{
-	"id":            reflect.Int,
-	"user_id":       reflect.Int,
+	"id":            reflect.Int64,
+	"user_id":       reflect.Int64,
 	"privacy":       reflect.Map, // JSONB
 	"notifications": reflect.Map, // JSONB
 	"language":      reflect.String,
@@ -48,8 +48,8 @@ var UserSettingsSchema = map[string]reflect.Kind{
 
 // SessionsCache
 var SessionsSchema = map[string]reflect.Kind{
-	"id":             reflect.Int,
-	"user_id":        reflect.Int,
+	"id":             reflect.Int64,
+	"user_id":        reflect.Int64,
 	"master_token":   reflect.String,
 	"device_token":   reflect.String, // FCM token
 	"device_info":    reflect.Map,    // JSONB
@@ -64,9 +64,9 @@ var SessionsSchema = map[string]reflect.Kind{
 
 // RelationsCache
 var RelationsSchema = map[string]reflect.Kind{
-	"id":           reflect.Int,
-	"primary_id":   reflect.Int,
-	"secondary_id": reflect.Int,
+	"id":           reflect.Int64,
+	"primary_id":   reflect.Int64,
+	"secondary_id": reflect.Int64,
 	"state":        reflect.Int,
 	"created_at":   reflect.Struct,
 	"updated_at":   reflect.Struct,
@@ -74,8 +74,8 @@ var RelationsSchema = map[string]reflect.Kind{
 
 // PostsCache
 var PostsSchema = map[string]reflect.Kind{
-	"id":         reflect.Int,
-	"user_id":    reflect.Int,
+	"id":         reflect.Int64,
+	"user_id":    reflect.Int64,
 	"content":    reflect.String,
 	"media_ids":  reflect.Slice, // UUID[]
 	"visibility": reflect.Int,
@@ -86,9 +86,9 @@ var PostsSchema = map[string]reflect.Kind{
 
 // CommentsCache
 var CommentsSchema = map[string]reflect.Kind{
-	"id":         reflect.Int,
-	"post_id":    reflect.Int,
-	"user_id":    reflect.Int,
+	"id":         reflect.Int64,
+	"post_id":    reflect.Int64,
+	"user_id":    reflect.Int64,
 	"content":    reflect.String,
 	"visibility": reflect.Bool,
 	"created_at": reflect.Struct,
@@ -97,17 +97,17 @@ var CommentsSchema = map[string]reflect.Kind{
 
 // LikesCache
 var LikesSchema = map[string]reflect.Kind{
-	"id":          reflect.Int,
+	"id":          reflect.Int64,
 	"target_type": reflect.Int,
-	"target_id":   reflect.Int,
-	"user_id":     reflect.Int,
+	"target_id":   reflect.Int64,
+	"user_id":     reflect.Int64,
 	"created_at":  reflect.Struct,
 }
 
 // MediaCache
 var MediaSchema = map[string]reflect.Kind{
-	"id":           reflect.Int,
-	"owner_id":     reflect.Int,
+	"id":           reflect.Int64,
+	"owner_id":     reflect.Int64,
 	"storage_path": reflect.String,
 	"visibility":   reflect.Bool,
 	"created_at":   reflect.Struct,
@@ -116,11 +116,11 @@ var MediaSchema = map[string]reflect.Kind{
 
 // ConversationsCache
 var ConversationsSchema = map[string]reflect.Kind{
-	"id":                          reflect.Int,
+	"id":                          reflect.Int64,
 	"type":                        reflect.Int,
 	"title":                       reflect.String,
-	"last_message_id":             reflect.Int,
-	"last_read_by_all_message_id": reflect.Int,
+	"last_message_id":             reflect.Int64,
+	"last_read_by_all_message_id": reflect.Int64,
 	"state":                       reflect.Int,
 	"created_at":                  reflect.Struct,
 	"updated_at":                  reflect.Struct,
@@ -128,9 +128,9 @@ var ConversationsSchema = map[string]reflect.Kind{
 
 // MembersCache
 var MembersSchema = map[string]reflect.Kind{
-	"id":              reflect.Int,
-	"conversation_id": reflect.Int,
-	"user_id":         reflect.Int,
+	"id":              reflect.Int64,
+	"conversation_id": reflect.Int64,
+	"user_id":         reflect.Int64,
 	"role":            reflect.Int,
 	"joined_at":       reflect.Struct,
 	"unread_count":    reflect.Int,
@@ -140,9 +140,9 @@ var MembersSchema = map[string]reflect.Kind{
 
 // MessagesCache
 var MessagesSchema = map[string]reflect.Kind{
-	"id":              reflect.Int,
-	"conversation_id": reflect.Int,
-	"sender_id":       reflect.Int,
+	"id":              reflect.Int64,
+	"conversation_id": reflect.Int64,
+	"sender_id":       reflect.Int64,
 	"message_type":    reflect.Int,
 	"state":           reflect.Int,
 	"content":         reflect.String,
